requirements.txt
alembic==1.13.2
async-timeout==4.0.3
asyncpg==0.29.0
greenlet==3.0.3
Mako==1.3.5
MarkupSafe==2.1.5
python-dotenv==1.0.1
SQLAlchemy==2.0.31
typing_extensions==4.12.2
